+++ src/parsing/tokenization/tokenization.c
@@ -12,7 +12,7 @@
 
 #include "parsing.h"
 
-static int	loop_content(t_input **input, t_env **env, char *line, int *i)
+static int	loop_content(t_input **input, t_env **env, char *line, size_t *i)
 {
 	while (line[*i] && ((line[*i] >= '\t' && line[*i] <= '\r')
 			|| line[*i] == ' '))
@@ -42,7 +42,7 @@ static int	loop_content(t_input **input, t_env **env, char *line, int *i)
 
 int	tokenization(t_input **input, t_env **env, char *line)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (check_quote(line))
